victoria-store-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2


PyJWT==2.8.0

